requirements.txt
crewai==0.45.0
langchain==0.1.16
langchain-openai==0.1.6
langchain-google-genai==1.0.5
streamlit==1.33.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.6.4
pyyaml==6.0.1
openai==1.23.6
google-generativeai==0.5.0
